ISmartContractTransactionService abstraction
