+++ src/Stratis.Bitcoin.Features.SmartContracts.Tests/Controllers/SmartContractWalletControllerTest.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using CSharpFunctionalExtensions;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -12,11 +11,8 @@ using Stratis.Bitcoin.Features.SmartContracts.Networks;
 using Stratis.Bitcoin.Features.SmartContracts.ReflectionExecutor.Consensus.Rules;
 using Stratis.Bitcoin.Features.SmartContracts.Wallet;
 using Stratis.Bitcoin.Features.Wallet;
-using Stratis.Bitcoin.Features.Wallet.Controllers;
 using Stratis.Bitcoin.Features.Wallet.Interfaces;
-using Stratis.Bitcoin.Features.Wallet.Models;
 using Stratis.Bitcoin.Tests.Wallet.Common;
-using Stratis.Bitcoin.Utilities;
 using Stratis.SmartContracts;
 using Stratis.SmartContracts.Core.Receipts;
 using Stratis.SmartContracts.Executor.Reflection;
@@ -34,6 +30,7 @@ namespace Stratis.Bitcoin.Features.SmartContracts.Tests.Controllers
         private readonly Network network;
         private readonly Mock<IReceiptRepository> receiptRepository;
         private readonly Mock<IWalletManager> walletManager;
+        private Mock<ISmartContractTransactionService> smartContractTransactionService;
 
         public SmartContractWalletControllerTest()
         {
@@ -44,6 +41,7 @@ namespace Stratis.Bitcoin.Features.SmartContracts.Tests.Controllers
             this.network = new SmartContractsRegTest();
             this.receiptRepository = new Mock<IReceiptRepository>();
             this.walletManager = new Mock<IWalletManager>();
+            this.smartContractTransactionService = new Mock<ISmartContractTransactionService>();
         }
 
         [Fact]
@@ -105,7 +103,8 @@ namespace Stratis.Bitcoin.Features.SmartContracts.Tests.Controllers
                 this.loggerFactory.Object,
                 this.network,
                 this.receiptRepository.Object,
-                this.walletManager.Object);
+                this.walletManager.Object,
+                this.smartContractTransactionService.Object);
 
             IActionResult result = controller.GetHistory(walletName, address.Address);
 
@@ -120,14 +119,14 @@ namespace Stratis.Bitcoin.Features.SmartContracts.Tests.Controllers
             ContractTransactionItem resultingCreate = model.ElementAt(0);
             Assert.Equal(ContractTransactionItemType.ContractCreate, resultingCreate.Type);
             Assert.Equal(createTransaction.SpendingDetails.TransactionId, resultingCreate.Hash);
-            Assert.Equal(createTransaction.SpendingDetails.Payments.First().Amount.ToUnit(MoneyUnit.BTC), resultingCreate.Amount);
+            Assert.Equal(createTransaction.SpendingDetails.Payments.First().Amount.ToUnit(MoneyUnit.Satoshi), resultingCreate.Amount);
             Assert.Equal(uint160.Zero.ToBase58Address(this.network), resultingCreate.To);
             Assert.Equal(createTransaction.SpendingDetails.BlockHeight, resultingCreate.BlockHeight);
 
             Assert.Equal(ContractTransactionItemType.Received, resultingTransaction.Type);
             Assert.Equal(address.Address, resultingTransaction.To);
             Assert.Equal(normalTransaction.Id, resultingTransaction.Hash);
-            Assert.Equal(normalTransaction.Amount.ToUnit(MoneyUnit.BTC), resultingTransaction.Amount);
+            Assert.Equal(normalTransaction.Amount.ToUnit(MoneyUnit.Satoshi), resultingTransaction.Amount);
             Assert.Equal(1, resultingTransaction.BlockHeight);
         }
     }
