+++ src/Stratis.Bitcoin.Features.SmartContracts/Wallet/ISmartContractTransactionService.cs
@@ -1,0 +1,10 @@
+﻿using Stratis.Bitcoin.Features.SmartContracts.Models;
+
+namespace Stratis.Bitcoin.Features.SmartContracts.Wallet
+{
+    public interface ISmartContractTransactionService
+    {
+        BuildCallContractTransactionResponse BuildCallTx(BuildCallContractTransactionRequest request);
+        BuildCreateContractTransactionResponse BuildCreateTx(BuildCreateContractTransactionRequest request);
+    }
+}
